@@ -2,41 +2,40 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <StartupObject>Rosalind.Core.Program</StartupObject>
+    <TargetFramework>net6.0</TargetFramework>
+    <StartupObject>Rosalind.Core.Program.Main</StartupObject>
+    <LangVersion>10</LangVersion>
+    <GenerateDocumentationFile>false</GenerateDocumentationFile>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="BooruSharp" Version="3.0.3" />
-    <PackageReference Include="Discord.Net.Labs.Commands" Version="3.1.6" />
-    <PackageReference Include="Discord.Net.Labs.Core" Version="3.1.6" />
-    <PackageReference Include="Discord.Net.Labs.Rest" Version="3.1.6" />
-    <PackageReference Include="Discord.Net.Labs.WebSocket" Version="3.1.6" />
-    <PackageReference Include="FluentArgs" Version="0.9.7" />
-    <PackageReference Include="log4net" Version="2.0.12" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="3.11.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.2" />
-    <PackageReference Include="MySql.Data" Version="8.0.26" />
+    <PackageReference Include="Discord.Net.Labs.Commands" Version="3.6.1" />
+    <PackageReference Include="Discord.Net.Labs.Core" Version="3.6.1" />
+    <PackageReference Include="Discord.Net.Labs.Rest" Version="3.6.1" />
+    <PackageReference Include="Discord.Net.Labs.WebSocket" Version="3.6.1" />
+    <PackageReference Include="log4net" Version="2.0.13" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="4.0.1" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
+    <PackageReference Include="MySql.Data" Version="8.0.27" />
     <PackageReference Include="SqlKata" Version="2.3.7" />
     <PackageReference Include="SqlKata.Execution" Version="2.3.7" />
+    <PackageReference Include="System.CommandLine" Version="2.0.0-beta2.21617.1" />
+    <PackageReference Include="System.CommandLine.DragonFruit" Version="0.4.0-alpha.21617.1" />
     <PackageReference Include="System.Data.SqlClient" Version="4.8.3" />
-    <PackageReference Include="System.Runtime.Caching" Version="5.0.0" />
-    <PackageReference Include="Victoria" Version="5.2.3" />
+    <PackageReference Include="System.Runtime.Caching" Version="6.0.0" />
+    <PackageReference Include="Victoria" Version="5.2.5" />
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="Lavalink.example.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
     <None Update="Lavalink.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Log4net.config">
+    <None Update="log4net.config">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
     <None Update="Setting.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-
 </Project>
