@@ -2,22 +2,21 @@
 using Rosalind.Core.Services;
 using System.Threading.Tasks;
 
-namespace Rosalind.Core.Commands.Game
+namespace Rosalind.Core.Commands.Game;
+
+public class Bank : ModuleBase<SocketCommandContext>
 {
-    public class Bank : ModuleBase<SocketCommandContext>
+    private readonly SqlService _sql;
+
+    public Bank(SqlService sql)
     {
-        private readonly SqlService _sql;
+        _sql = sql;
+    }
 
-        public Bank(SqlService sql)
-        {
-            _sql = sql;
-        }
-
-        [Command("은행")]
-        public async Task BankAsync()
-        {
-            var user = _sql.GetUser(Context.Guild.Id, Context.User.Id);
-            await ReplyAsync($"💰 {Context.User.Username}님은 현재 `{string.Format("{0:#,0}", user.Coin)}` 코인을 소지하고 있습니다.");
-        }
+    [Command("은행")]
+    public async Task BankAsync()
+    {
+        var user = _sql.GetUser(Context.Guild.Id, Context.User.Id);
+        await ReplyAsync($"💰 {Context.User.Username}님은 현재 `{user.Coin:#,0}` 코인을 소지하고 있습니다.");
     }
 }
